+++ apps/docs/src/content/03-components/form-controls/date-picker/examples/fieldDescription.tsx
@@ -1,0 +1,10 @@
+import {
+  DatePicker,
+  FieldDescription,
+  Label,
+} from "@mittwald/flow-react-components";
+
+<DatePicker>
+  <Label>Datum</Label>
+  <FieldDescription>Weitere Informationen</FieldDescription>
+</DatePicker>;