+++ apps/docs/src/content/03-components/form-controls/date-range-picker/examples/fieldDescription.tsx
@@ -1,0 +1,10 @@
+import {
+  DateRangePicker,
+  FieldDescription,
+  Label,
+} from "@mittwald/flow-react-components";
+
+<DateRangePicker>
+  <Label>Zeitraum</Label>
+  <FieldDescription>Weitere Informationen</FieldDescription>
+</DateRangePicker>;