+++ apps/docs/src/content/03-components/form-controls/search-field/examples/align.tsx
@@ -6,5 +6,5 @@ import {
 
 <Align>
   <SearchField />
-  <Button>Hinzufügen</Button>
+  <Button>Suchen</Button>
 </Align>;