+++ apps/docs/src/content/03-components/form-controls/text-area/examples/form.tsx
@@ -1,0 +1,35 @@
+import {
+  Button,
+  Label,
+  Section,
+  TextArea,
+} from "@mittwald/flow-react-components";
+import { useForm } from "react-hook-form";
+import {
+  Form,
+  typedField,
+} from "@mittwald/flow-react-components/react-hook-form";
+import { sleep } from "@/content/03-components/actions/action/examples/lib";
+
+export default () => {
+  const form = useForm<{ message: string }>();
+  const Field = typedField(form);
+
+  return (
+    <Section>
+      <Form form={form} onSubmit={sleep}>
+        <Field
+          name="message"
+          rules={{
+            required: "Bitte gib eine Nachricht ein",
+          }}
+        >
+          <TextArea>
+            <Label>Nachricht</Label>
+          </TextArea>
+        </Field>
+        <Button type="submit">Senden</Button>
+      </Form>
+    </Section>
+  );
+};