+++ packages/components/src/components/ChartTooltip/index.ts
@@ -1,0 +1,4 @@
+import { ChartTooltip } from "./ChartTooltip";
+
+export { type ChartTooltipProps, ChartTooltip } from "./ChartTooltip";
+export default ChartTooltip;