+++ packages/components/src/components/Flex/Flex.tsx
@@ -20,6 +20,16 @@ export interface FlexProps extends PropsWithChildren, PropsWithClassName {
   grow?: boolean;
   /** The flexWrap value of the element. @default "nowrap" */
   wrap?: CSSProperties["flexWrap"];
+  /** The padding of the element. */
+  padding?: "xs" | "s" | "m" | "l" | "xl";
+  /** The padding top of the element. */
+  paddingTop?: "xs" | "s" | "m" | "l" | "xl";
+  /** The padding bottom of the element. */
+  paddingBottom?: "xs" | "s" | "m" | "l" | "xl";
+  /** The padding left of the element. */
+  paddingLeft?: "xs" | "s" | "m" | "l" | "xl";
+  /** The padding right of the element. */
+  paddingRight?: "xs" | "s" | "m" | "l" | "xl";
 }
 
 /** @flr-generate all */
@@ -35,6 +45,11 @@ export const Flex: FC<FlexProps> = (props) => {
     rowGap,
     grow,
     wrap = "nowrap",
+    padding,
+    paddingTop,
+    paddingBottom,
+    paddingLeft,
+    paddingRight,
   } = props;
 
   const rootClassName = clsx(styles.flex, className);
@@ -69,6 +84,26 @@ export const Flex: FC<FlexProps> = (props) => {
             : undefined,
         flexGrow: grow ? 1 : undefined,
         flexWrap: wrap,
+        paddingTop: paddingTop
+          ? `var(--size-px--${paddingTop})`
+          : padding
+            ? `var(--size-px--${padding})`
+            : undefined,
+        paddingBottom: paddingBottom
+          ? `var(--size-px--${paddingBottom})`
+          : padding
+            ? `var(--size-px--${padding})`
+            : undefined,
+        paddingLeft: paddingLeft
+          ? `var(--size-px--${paddingLeft})`
+          : padding
+            ? `var(--size-px--${padding})`
+            : undefined,
+        paddingRight: paddingRight
+          ? `var(--size-px--${paddingRight})`
+          : padding
+            ? `var(--size-px--${padding})`
+            : undefined,
       }}
     >
       {children}