+++ packages/components/src/components/Legend/view.ts
@@ -1,0 +1,10 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import type { Legend } from "./Legend";
+import type { ViewComponent } from "@/lib/viewComponentContext";
+
+declare global {
+  interface FlowViewComponents {
+    Legend: ViewComponent<typeof Legend>;
+  }
+}