+++ packages/components/src/views/CartesianGridView.tsx
@@ -1,0 +1,16 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+import React, { type FC, useContext } from "react";
+import {
+  CartesianGrid,
+  type CartesianGridProps,
+} from "@/components/CartesianGrid";
+import { viewComponentContext } from "@/lib/viewComponentContext/viewComponentContext";
+
+const CartesianGridView: FC<CartesianGridProps> = (props) => {
+  const View =
+    useContext(viewComponentContext)["CartesianGrid"] ?? CartesianGrid;
+  return <View {...props} />;
+};
+
+export default CartesianGridView;