+++ packages/ext-bridge/src/react/hooks/useExtBridge.ts
@@ -2,6 +2,6 @@ import { usePromise } from "@mittwald/react-use-promise";
 
 export const useExtBridge = () =>
   usePromise(async () => {
-    await globalThis.mwExtBridge.ready;
-    return globalThis.mwExtBridge;
+    await globalThis.mittwald.extBridge.ready;
+    return globalThis.mittwald.extBridge;
   }, []);