+++ packages/remote-react-components/src/auto-generated/Legend.ts
@@ -1,0 +1,21 @@
+/* prettier-ignore */
+/* This file is auto-generated with the remote-components-generator */
+"use client";
+import createFlowRemoteComponent from "@/components/createFlowRemoteComponent";
+import { RemoteLegendElement } from "@mittwald/flow-remote-elements";
+export { type RemoteLegendElement } from "@mittwald/flow-remote-elements";
+
+export const Legend = createFlowRemoteComponent(
+  "flr-legend",
+  "Legend",
+  {
+    clearPropsContext: false,
+  },
+  RemoteLegendElement,
+  {
+    slotProps: {
+      wrapper: false,
+    },
+    eventProps: {},
+  },
+);